refactor(Color): 컬러 패키지 이동
