+++ src/main/java/chess/domain/board/InitialBoardGenerator.java
@@ -8,7 +8,7 @@ import java.util.function.Function;
 import java.util.stream.IntStream;
 
 import chess.domain.piece.Bishop;
-import chess.domain.piece.Color;
+import chess.domain.Color;
 import chess.domain.piece.Empty;
 import chess.domain.piece.King;
 import chess.domain.piece.Knight;
