+++ src/main/java/chess/domain/piece/move/pawn/PawnSupport.java
@@ -2,7 +2,7 @@ package chess.domain.piece.move.pawn;
 
 import chess.domain.board.LineNumber;
 import chess.domain.board.Point;
-import chess.domain.piece.Color;
+import chess.domain.Color;
 
 public class PawnSupport {
