+++ src/test/java/chess/domain/board/BoardTest.java
@@ -8,7 +8,7 @@ import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import chess.domain.piece.Color;
+import chess.domain.Color;
 import chess.domain.piece.King;
 import chess.domain.piece.Queen;
