+++ src/test/java/chess/domain/game/RunningTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 
 import chess.domain.board.BoardFixtures;
 import chess.domain.board.Point;
-import chess.domain.piece.Color;
+import chess.domain.Color;
 import chess.domain.piece.King;
 import chess.domain.piece.Queen;
 import chess.dto.BoardResponse;
